+++ src/getStepStartStates.test.ts
@@ -83,6 +83,7 @@ import {
 import {
   STEP_RM_REDIS_CACHES,
   STEP_RM_REDIS_FIREWALL_RULES,
+  STEP_RM_REDIS_LINKED_SERVERS,
 } from './steps/resource-manager/redis-cache';
 import {
   STEP_RM_EVENT_GRID_DOMAINS,
@@ -159,6 +160,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_BATCH_CERTIFICATE]: { disabled: true },
       [STEP_RM_REDIS_CACHES]: { disabled: true },
       [STEP_RM_REDIS_FIREWALL_RULES]: { disabled: true },
+      [STEP_RM_REDIS_LINKED_SERVERS]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAINS]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAIN_TOPICS]: { disabled: true },
       [STEP_RM_EVENT_GRID_TOPICS]: { disabled: true },
@@ -228,6 +230,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_BATCH_CERTIFICATE]: { disabled: true },
       [STEP_RM_REDIS_CACHES]: { disabled: true },
       [STEP_RM_REDIS_FIREWALL_RULES]: { disabled: true },
+      [STEP_RM_REDIS_LINKED_SERVERS]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAINS]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAIN_TOPICS]: { disabled: true },
       [STEP_RM_EVENT_GRID_TOPICS]: { disabled: true },
@@ -297,6 +300,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_BATCH_CERTIFICATE]: { disabled: false },
       [STEP_RM_REDIS_CACHES]: { disabled: false },
       [STEP_RM_REDIS_FIREWALL_RULES]: { disabled: false },
+      [STEP_RM_REDIS_LINKED_SERVERS]: { disabled: false },
       [STEP_RM_EVENT_GRID_DOMAINS]: { disabled: false },
       [STEP_RM_EVENT_GRID_DOMAIN_TOPICS]: { disabled: false },
       [STEP_RM_EVENT_GRID_TOPICS]: { disabled: false },