+++ src/steps/resource-manager/redis-cache/client.test.ts
@@ -7,6 +7,7 @@ import { IntegrationConfig } from '../../../types';
 import { RedisCacheClient } from './client';
 import {
   RedisFirewallRule,
+  RedisLinkedServerProperties,
   RedisResource,
 } from '@azure/arm-rediscache/esm/models';
 
@@ -25,10 +26,10 @@ afterEach(async () => {
 });
 
 describe('iterate caches', () => {
-  test('all', async () => {
+  test('j1dev - resource group', async () => {
     recording = setupAzureRecording({
       directory: __dirname,
-      name: 'iterateCaches',
+      name: 'iterateCaches-j1dev',
     });
 
     const client = new RedisCacheClient(
@@ -66,6 +67,107 @@ describe('iterate caches', () => {
           'maxfragmentationmemory-reserved': expect.any(String),
           'maxmemory-delta': expect.any(String),
           'maxmemory-reserved': expect.any(String),
+          'maxmemory-policy': expect.any(String),
+        },
+        redisVersion: expect.any(String),
+        sku: {
+          capacity: expect.any(Number),
+          family: expect.any(String),
+          name: expect.any(String),
+        },
+        sslPort: expect.any(Number),
+        tags: {},
+        type: 'Microsoft.Cache/Redis',
+      }),
+    );
+
+    expect(resources).toContainEqual(
+      expect.objectContaining({
+        accessKeys: null,
+        enableNonSslPort: expect.any(Boolean),
+        hostName: 'keionned-j1dev-primary-redis-cache.redis.cache.windows.net',
+        id: `/subscriptions/${config.subscriptionId}/resourceGroups/${resourceGroupInfo.resourceGroupName}/providers/Microsoft.Cache/Redis/keionned-j1dev-primary-redis-cache`,
+        instances: [
+          { isMaster: expect.any(Boolean), sslPort: expect.any(Number) },
+          { isMaster: expect.any(Boolean), sslPort: expect.any(Number) },
+        ],
+        linkedServers: [
+          {
+            id: `/subscriptions/${config.subscriptionId}/resourceGroups/${resourceGroupInfo.resourceGroupName}/providers/Microsoft.Cache/Redis/keionned-j1dev-primary-redis-cache/linkedServers/keionned-j1dev-secondary-redis-cache`,
+          },
+        ],
+        location: 'East US',
+        minimumTlsVersion: expect.any(String),
+        name: 'keionned-j1dev-primary-redis-cache',
+        port: expect.any(Number),
+        provisioningState: 'Succeeded',
+        redisConfiguration: {
+          maxclients: expect.any(String),
+          'maxfragmentationmemory-reserved': expect.any(String),
+          'maxmemory-delta': expect.any(String),
+          'maxmemory-reserved': expect.any(String),
+          'maxmemory-policy': expect.any(String),
+        },
+        redisVersion: expect.any(String),
+        sku: {
+          capacity: expect.any(Number),
+          family: expect.any(String),
+          name: expect.any(String),
+        },
+        sslPort: expect.any(Number),
+        tags: {},
+        type: 'Microsoft.Cache/Redis',
+      }),
+    );
+  });
+
+  test('j1dev-secondary-redis-cache-resource-group - resource group', async () => {
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: 'iterateCaches-j1dev-secondary-redis-cache-resource-group',
+    });
+
+    const client = new RedisCacheClient(
+      config,
+      createMockIntegrationLogger(),
+      true,
+    );
+    const resources: RedisResource[] = [];
+    const resourceGroupInfo = {
+      resourceGroupName: 'j1dev-secondary-redis-cache-resource-group',
+    };
+
+    await client.iterateCaches(resourceGroupInfo, (e) => {
+      resources.push(e);
+    });
+
+    expect(resources).toContainEqual(
+      expect.objectContaining({
+        accessKeys: null,
+        enableNonSslPort: expect.any(Boolean),
+        hostName:
+          'keionned-j1dev-secondary-redis-cache.redis.cache.windows.net',
+        id: `/subscriptions/${config.subscriptionId}/resourceGroups/${resourceGroupInfo.resourceGroupName}/providers/Microsoft.Cache/Redis/keionned-j1dev-secondary-redis-cache`,
+        instances: [
+          { isMaster: expect.any(Boolean), sslPort: expect.any(Number) },
+          { isMaster: expect.any(Boolean), sslPort: expect.any(Number) },
+        ],
+        linkedServers: [
+          {
+            id: `/subscriptions/${config.subscriptionId}/resourceGroups/${resourceGroupInfo.resourceGroupName}/providers/Microsoft.Cache/Redis/keionned-j1dev-secondary-redis-cache/linkedServers/keionned-j1dev-primary-redis-cache`,
+          },
+        ],
+        location: 'West US',
+        minimumTlsVersion: expect.any(String),
+        name: 'keionned-j1dev-secondary-redis-cache',
+        port: expect.any(Number),
+        provisioningState: 'Succeeded',
+        redisConfiguration: {
+          maxclients: expect.any(String),
+          'maxfragmentationmemory-reserved': expect.any(String),
+          'maxmemory-delta': expect.any(String),
+          'maxmemory-reserved': expect.any(String),
+          'maxmemory-policy': expect.any(String),
         },
         redisVersion: expect.any(String),
         sku: {
@@ -110,6 +212,77 @@ describe('iterate firewall rules', () => {
         startIP: '1.2.3.4',
         endIP: '2.3.4.5',
         type: 'Microsoft.Cache/Redis/firewallRules',
+      }),
+    );
+  });
+});
+
+describe('iterate linked servers', () => {
+  test('j1dev - resource group', async () => {
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: 'iterateLinkedServers-j1dev',
+    });
+
+    const client = new RedisCacheClient(
+      config,
+      createMockIntegrationLogger(),
+      true,
+    );
+    const resources: RedisLinkedServerProperties[] = [];
+    const redisCacheInfo = {
+      resourceGroupName: 'j1dev',
+      redisCacheName: 'keionned-j1dev-primary-redis-cache',
+    };
+
+    await client.iterateLinkedServers(redisCacheInfo, (e) => {
+      resources.push(e);
+    });
+
+    expect(resources).toContainEqual(
+      expect.objectContaining({
+        id: `/subscriptions/${config.subscriptionId}/resourceGroups/${redisCacheInfo.resourceGroupName}/providers/Microsoft.Cache/Redis/keionned-j1dev-primary-redis-cache/linkedServers/keionned-j1dev-secondary-redis-cache`,
+        name: 'keionned-j1dev-secondary-redis-cache',
+        linkedRedisCacheId:
+          '/subscriptions/40474ebe-55a2-4071-8fa8-b610acdd8e56/resourceGroups/j1dev-secondary-redis-cache-resource-group/providers/Microsoft.Cache/Redis/keionned-j1dev-secondary-redis-cache',
+        linkedRedisCacheLocation: 'West US',
+        provisioningState: 'Succeeded',
+        serverRole: 'Secondary',
+        type: 'Microsoft.Cache/Redis/linkedServers',
+      }),
+    );
+  });
+
+  test('j1dev-secondary-redis-cache-resource-group - resource group', async () => {
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: 'iterateLinkedServers-j1dev-secondary-redis-cache-resource-group',
+    });
+
+    const client = new RedisCacheClient(
+      config,
+      createMockIntegrationLogger(),
+      true,
+    );
+    const resources: RedisLinkedServerProperties[] = [];
+    const redisCacheInfo = {
+      resourceGroupName: 'j1dev-secondary-redis-cache-resource-group',
+      redisCacheName: 'keionned-j1dev-secondary-redis-cache',
+    };
+
+    await client.iterateLinkedServers(redisCacheInfo, (e) => {
+      resources.push(e);
+    });
+
+    expect(resources).toContainEqual(
+      expect.objectContaining({
+        id: `/subscriptions/${config.subscriptionId}/resourceGroups/${redisCacheInfo.resourceGroupName}/providers/Microsoft.Cache/Redis/keionned-j1dev-secondary-redis-cache/linkedServers/keionned-j1dev-primary-redis-cache`,
+        name: 'keionned-j1dev-primary-redis-cache',
+        linkedRedisCacheId: `/subscriptions/${config.subscriptionId}/resourceGroups/j1dev/providers/Microsoft.Cache/Redis/keionned-j1dev-primary-redis-cache`,
+        linkedRedisCacheLocation: 'East US',
+        provisioningState: 'Succeeded',
+        serverRole: 'Primary',
+        type: 'Microsoft.Cache/Redis/linkedServers',
       }),
     );
   });