+++ src/steps/resource-manager/redis-cache/client.ts
@@ -2,6 +2,7 @@ import { RedisManagementClient } from '@azure/arm-rediscache';
 import {
   RedisResource,
   RedisFirewallRule,
+  RedisLinkedServerWithProperties,
 } from '@azure/arm-rediscache/esm/models';
 import {
   Client,
@@ -37,7 +38,7 @@ export class RedisCacheClient extends Client {
   }
 
   /**
-   * Retrieves all Redis Firewall Rules for a Redis Caches in a Resource Group for an Azure Subscription
+   * Retrieves all Redis Firewall Rules for a Redis Cache in a Resource Group for an Azure Subscription
    * @param redisCacheInfo An object containing information about the Redis Cache needed to retrieve the Redis Firewall Rules. This should include the Resource Group name that the Redis Cache belongs to and the name of the Redis Cache.
    * @param callback A callback function to be called after retrieving Redis Firewall Rules
    */
@@ -62,7 +63,35 @@ export class RedisCacheClient extends Client {
           ),
         listNext: serviceClient.firewallRules.listByRedisResourceNext,
       } as ListResourcesEndpoint,
-      resourceDescription: 'redisCache.firewallRules',
+      resourceDescription: 'redisCache.firewallRule',
+      callback,
+    });
+  }
+
+  /**
+   * Retrieves all Redis Linked Servers for a Redis Cache in a Resource Group for an Azure Subscription
+   * @param redisCacheInfo An object containing information about the Redis Cache needed to retrieve the Redis Linked Server. This should include the Resource Group name that the Redis Cache belongs to and the name of the Redis Cache.
+   * @param callback A callback function to be called after retrieving the Redis Linked Server
+   */
+  public async iterateLinkedServers(
+    redisCacheInfo: { resourceGroupName: string; redisCacheName: string },
+    callback: (s: RedisLinkedServerWithProperties) => void | Promise<void>,
+  ): Promise<void> {
+    const serviceClient = await this.getAuthenticatedServiceClient(
+      RedisManagementClient,
+    );
+
+    const { resourceGroupName, redisCacheName } = redisCacheInfo;
+
+    return iterateAllResources({
+      logger: this.logger,
+      serviceClient,
+      resourceEndpoint: {
+        list: async () =>
+          serviceClient.linkedServer.list(resourceGroupName, redisCacheName),
+        listNext: serviceClient.linkedServer.listNext,
+      } as ListResourcesEndpoint,
+      resourceDescription: 'redisCache.linkedServer',
       callback,
     });
   }