+++ src/steps/resource-manager/redis-cache/index.ts
@@ -17,14 +17,18 @@ import createResourceGroupResourceRelationship from '../utils/createResourceGrou
 import {
   createRedisCacheEntity,
   createRedisFirewallRuleEntity,
+  createRedisLinkedServerRelationshipProperties,
 } from './converters';
 import {
   RedisCacheEntities,
   RedisCacheRelationships,
   STEP_RM_REDIS_CACHES,
   STEP_RM_REDIS_FIREWALL_RULES,
+  STEP_RM_REDIS_LINKED_SERVERS,
 } from './constants';
 import { resourceGroupName } from '../../../azure/utils';
+import { RedisLinkedServerWithProperties } from '@azure/arm-rediscache/esm/models';
+const SECONDARY_SERVER_ROLE = 'Secondary';
 export * from './constants';
 
 export async function fetchRedisCaches(
@@ -94,6 +98,54 @@ export async function fetchRedisFirewallRules(
   );
 }
 
+const cacheIsLinkedToSecondaryCache = (
+  linkedServer: RedisLinkedServerWithProperties,
+): Boolean => linkedServer.serverRole === SECONDARY_SERVER_ROLE;
+
+export async function fetchRedisLinkedServers(
+  executionContext: IntegrationStepContext,
+): Promise<void> {
+  const { instance, logger, jobState } = executionContext;
+  const accountEntity = await jobState.getData<Entity>(ACCOUNT_ENTITY_TYPE);
+  const webLinker = createAzureWebLinker(accountEntity.defaultDomain as string);
+  const client = new RedisCacheClient(instance.config, logger);
+
+  await jobState.iterateEntities(
+    { _type: RedisCacheEntities.CACHE._type },
+    async (redisCacheEntity) => {
+      const { id, name } = redisCacheEntity;
+      const resourceGroup = resourceGroupName(id, true)!;
+
+      await client.iterateLinkedServers(
+        { resourceGroupName: resourceGroup, redisCacheName: name as string },
+        async (linkedServer) => {
+          const { linkedRedisCacheId } = linkedServer;
+          if (cacheIsLinkedToSecondaryCache(linkedServer)) {
+            const secondaryCacheEntity = await jobState.findEntity(
+              linkedRedisCacheId,
+            );
+
+            if (secondaryCacheEntity) {
+              await jobState.addRelationship(
+                createDirectRelationship({
+                  _class: RelationshipClass.CONNECTS,
+                  from: redisCacheEntity,
+                  to: secondaryCacheEntity,
+                  properties: createRedisLinkedServerRelationshipProperties(
+                    webLinker,
+                    redisCacheEntity,
+                    linkedServer,
+                  ),
+                }),
+              );
+            }
+          }
+        },
+      );
+    },
+  );
+}
+
 export const redisCacheSteps: Step<
   IntegrationStepExecutionContext<IntegrationConfig>
 >[] = [
@@ -116,5 +168,19 @@ export const redisCacheSteps: Step<
       STEP_RM_REDIS_CACHES,
     ],
     executionHandler: fetchRedisFirewallRules,
+  },
+  {
+    id: STEP_RM_REDIS_LINKED_SERVERS,
+    name: 'Redis Linked Servers',
+    entities: [],
+    relationships: [
+      RedisCacheRelationships.REDIS_CACHE_IS_LINKED_TO_REDIS_CACHE,
+    ],
+    dependsOn: [
+      STEP_AD_ACCOUNT,
+      STEP_RM_RESOURCES_RESOURCE_GROUPS,
+      STEP_RM_REDIS_CACHES,
+    ],
+    executionHandler: fetchRedisLinkedServers,
   },
 ];